Move back posts to create flat directory structure
The initial plan was to create year specific directories so that we can
add dates to the posts. But because this may be messy in having to
specify dates, I think it'll be easier to have a flat directory
structure for all posts. Although this directory will get long, it'll be
easier to reference the title, rather than having to go through the year
first to access it for editing.